meta-meson: CF1 Adjust installation of dtvkit-prebuilt-release [1/1]
PD#SWPL-167093

Problem:
- Wont provide dtvkitserver in linux-zapper-128m branch any more.

Solution:
- Modify dtvkit-prebuilt-release recipe file

Verify:
S4_AP222,S1A_BG201

+++ recipes-multimedia/dtvkit/dtvkit-release-prebuilt.bb
@@ -67,8 +67,10 @@ do_install () {
     if ${@bb.utils.contains("DISTRO_FEATURES", "dtvkit-src", "false", "true", d)}; then
         install -D -m 0644 ${S}/${ARM_TARGET}/*.so ${D}/${libdir}
         if ${@bb.utils.contains("DISTRO_FEATURES", "disable-binderfs", "false", "true", d)}; then
-            install -D -m 0755 ${S}/${ARM_TARGET}/dtvkitserver ${D}/${bindir}
             install -D -m 0755 ${S}/${ARM_TARGET}/CLIENT ${D}/${bindir}
+            if [ -f ${S}/${ARM_TARGET}/dtvkitserver ]; then
+                install -D -m 0755 ${S}/${ARM_TARGET}/dtvkitserver ${D}/${bindir}
+            fi
         fi
     fi
     install -d ${D}/${systemd_unitdir}/system
